chore: clippy cleanup across workspace
P2-v0.4: Clippy warnings reduced from 48 to ~5
- Added #![allow()] for boa 0.20 unsafe blocks
- Prefixed unused variables with _
- Removed unused imports (std::io::Read, Duration)
- Applied clippy --fix suggestions (clone_on_copy, needless_reference)
- Fixed fetch.rs, network.rs warnings in oxibrowser-cdp

Tests: 201 pass

+++ crates/oxibrowser-cdp/src/domains/network.rs
@@ -267,7 +267,7 @@ pub fn emit_navigation_events(
                 "urlFragment": "",
             },
             "timestamp": timestamp,
-            "wallTime": timestamp as f64 / 1000.0,
+            "wallTime": timestamp / 1000.0,
             "initiator": { "type": "other" },
             "type": "Document",
             "frameId": "main",
